refactor(cmd): extract SSH_KEY parsing, add tests
This extracts the SSH_KEY parsing code in cmd.ConfigSet into its own
function to make it easier to test and adds tests. It also changes the
order of evaluation to prefer the quickest and safest path.

The code also changes the private key regex to a stricter regex which
also matches new format OPENSSH keys and is required to support keys
of the ed25519 type.